Complexity comments
